Backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pillow==8.3.1
numpy==1.21.0
tensorflow==2.8.0
torch==1.9.0
torchvision==0.10.0
opencv-python==4.5.3.56
scikit-learn==1.0.0
nibabel==3.2.2
albumentations==1.1.0
python-dotenv==0.19.0 